move worker logic into worker cmd
