+++ golang/cmd/worker/worker_test.go
@@ -1,4 +1,4 @@
-package worker
+package main
 
 import (
 	"syscall"
